requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
pypdf
unstructured[pdf]
openai
langchain
langchain-community
chromadb
sentence-transformers
langchain-text-splitters
